Use all keys when fetching dataset from MongoDB
Since documents stored in a MongoDB collection are schemaless they can
vary per document in their set of keys.  Using only the first document's
keys can give an inaccurate set of columns for the data set, so this
change generates the union of all document keys in the fetched data
when specifying dataset columns.